Correcciones para cumplir historias de usuario.

+++ Web/eliminar-residencia.php
@@ -7,12 +7,18 @@
 	include('componentes/funciones-usuarios.php');
 	include('componentes/solo-admin.php');
 	$id = mysqli_escape_string($con, $_GET['id']);
+	$sql = mysqli_query($con, "SELECT * FROM semanas WHERE residencia=".$id." AND reservado=1");
+	if (mysqli_num_rows($sql) > 0) {
+		header('Location: index.php?error=3');
+		exit();
+	}
 	$query = "DELETE FROM `residencias` WHERE id=".$id;
 	$sql = mysqli_query($con, $query);
 	if ($sql) {
 		$sql = mysqli_query($con, "DELETE FROM `semanas` WHERE residencia=".$id);
 		$sql = mysqli_query($con, "DELETE FROM `subastas` WHERE residencia=".$id);
 		header('Location: index.php?exito=1');
+		exit();
 	}
 	else header('Location: index.php?error=2');
 ?>
